Add support for ml2 ovs driver
Neutron ML2 OVS driver uses iptables as default firewall driver.
In this case nova
1) adds VM's tap interface to linux bridge
2) create a veth pair between this linux bridge and ovs br-int
Then neutron programs iptables firewall rules on this bridge.
In the fake nova driver we need to the same, but use veth pair
for tap device. One end of veth pair is given to namespace and
the other to linux bridge.